@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FormsModule } from '@angular/forms';
@@ -11,24 +11,17 @@ import { HomeComponent } from './home/home.component';
 import { DetailsPopupComponent } from './details-popup/details-popup.component';
 import { BlogsComponent } from './blogs/blogs.component';
 import { AboutusComponent } from './aboutus/aboutus.component';
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    ProductsComponent,
-    RecipesListComponent,
-    HomeComponent,
-    DetailsPopupComponent,
-    BlogsComponent,
-    AboutusComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    FormsModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HeaderComponent,
+        ProductsComponent,
+        RecipesListComponent,
+        HomeComponent,
+        DetailsPopupComponent,
+        BlogsComponent,
+        AboutusComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        FormsModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
